Move runtime/constants to lib/genesis
