+++ lib/runtime/wasmer/exports_test.go
@@ -14,8 +14,8 @@ import (
 	"github.com/ChainSafe/gossamer/internal/log"
 	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/crypto/ed25519"
+	"github.com/ChainSafe/gossamer/lib/genesis"
 	"github.com/ChainSafe/gossamer/lib/runtime"
-	"github.com/ChainSafe/gossamer/lib/runtime/constants"
 	"github.com/ChainSafe/gossamer/lib/runtime/storage"
 	"github.com/ChainSafe/gossamer/lib/runtime/wasmer/testdata"
 	"github.com/ChainSafe/gossamer/lib/trie"
@@ -358,7 +358,7 @@ func TestInstance_GrandpaAuthorities_NodeRuntime(t *testing.T) {
 	value, err := common.HexToBytes("0x0108eea1eabcac7d2c8a6459b7322cf997874482bfc3d2ec7a80888a3a7d714103640100000000000000b64994460e59b30364cad3c92e3df6052f9b0ebbb8f88460c194dc5794d6d7170100000000000000") //nolint:lll
 	require.NoError(t, err)
 
-	key := common.MustHexToBytes(constants.GrandpaAuthoritiesKeyHex)
+	key := common.MustHexToBytes(genesis.GrandpaAuthoritiesKeyHex)
 	tt.Put(key, value)
 
 	rt := NewTestInstanceWithTrie(t, runtime.NODE_RUNTIME, tt)
@@ -386,7 +386,7 @@ func TestInstance_GrandpaAuthorities_PolkadotRuntime(t *testing.T) {
 	value, err := common.HexToBytes("0x0108eea1eabcac7d2c8a6459b7322cf997874482bfc3d2ec7a80888a3a7d714103640100000000000000b64994460e59b30364cad3c92e3df6052f9b0ebbb8f88460c194dc5794d6d7170100000000000000") //nolint:lll
 	require.NoError(t, err)
 
-	key := common.MustHexToBytes(constants.GrandpaAuthoritiesKeyHex)
+	key := common.MustHexToBytes(genesis.GrandpaAuthoritiesKeyHex)
 	tt.Put(key, value)
 
 	rt := NewTestInstanceWithTrie(t, runtime.POLKADOT_RUNTIME, tt)
@@ -449,13 +449,13 @@ func TestInstance_BabeConfiguration_NodeRuntime_WithAuthorities(t *testing.T) {
 
 	rvalue, err := common.HexToHash("0x01")
 	require.NoError(t, err)
-	key := common.MustHexToBytes(constants.BABERandomnessKeyHex)
+	key := common.MustHexToBytes(genesis.BABERandomnessKeyHex)
 	tt.Put(key, rvalue[:])
 
 	avalue, err := common.HexToBytes("0x08eea1eabcac7d2c8a6459b7322cf997874482bfc3d2ec7a80888a3a7d714103640100000000000000b64994460e59b30364cad3c92e3df6052f9b0ebbb8f88460c194dc5794d6d7170100000000000000") //nolint:lll
 	require.NoError(t, err)
 
-	key = common.MustHexToBytes(constants.BABEAuthoritiesKeyHex)
+	key = common.MustHexToBytes(genesis.BABEAuthoritiesKeyHex)
 	tt.Put(key, avalue)
 
 	rt := NewTestInstanceWithTrie(t, runtime.NODE_RUNTIME, tt)
